I have lost all sense of what is even happening
