+++ interp/src/errors.rs
@@ -1,7 +1,5 @@
+use crate::flatten::flat_ir::prelude::AssignedValue;
 use crate::values::Value;
-use crate::{
-    flatten::flat_ir::prelude::AssignedValue, utils::assignment_to_string,
-};
 use calyx_ir::{self as ir, Assignment, Id};
 use calyx_utils::Error as CalyxError;
 use rustyline::error::ReadlineError;
@@ -204,6 +202,15 @@ pub enum InterpreterError {
 
     #[error(transparent)]
     SerializationError(#[from] crate::serialization::SerializationError),
+}
+
+pub fn assignment_to_string(
+    assignment: &ir::Assignment<ir::Nothing>,
+) -> String {
+    let mut str = vec![];
+    ir::Printer::write_assignment(assignment, 0, &mut str)
+        .expect("Write Failed");
+    String::from_utf8(str).expect("Found invalid UTF-8")
 }
 
 impl InterpreterError {
